requirements.txt
ftfy 6.1.3
regex 2020.10.15
tqdm 4.50.2
torch 2.1.1
torchvision 0.16.1
torchvision
pillow 8.0.1
numpy 1.19.2